Big fixes with arguments and save
The current state was completely messed up with saving and all the
different scenarios.  This update brings back my main intent of the
code and also does a separate merge of the attributes for the use of
the FileObj.  It also adds Blob support, and I finally written some
tests.